+++ libraries/session/src/main/java/androidx/media3/session/MediaControllerImplLegacy.java
@@ -18,7 +18,6 @@ package androidx.media3.session;
 import static androidx.media3.session.MediaUtils.calculateBufferedPercentage;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
-import static com.google.common.base.Preconditions.checkState;
 import static java.lang.Math.max;
 import static java.lang.Math.min;
 import static java.lang.String.format;
@@ -115,6 +114,7 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
   private ControllerInfo controllerInfo;
   private long currentPositionMs;
   private long lastSetPlayWhenReadyCalledTimeMs;
+  private boolean hasInitializedLegacyPlaylist;
 
   public MediaControllerImplLegacy(
       Context context,
@@ -214,6 +214,7 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
         /* discontinuityReason= */ null,
         /* mediaItemTransitionReason= */ null);
 
+    hasInitializedLegacyPlaylist = false;
     controllerCompat.getTransportControls().stop();
   }
 
@@ -280,7 +281,7 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
         /* discontinuityReason= */ null,
         /* mediaItemTransitionReason= */ null);
 
-    if (hasMedia()) {
+    if (canInitializeLegacyPlaylist(controllerInfo)) {
       initializeLegacyPlaylist();
     }
   }
@@ -700,7 +701,7 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
         maskedControllerInfo,
         /* discontinuityReason= */ null,
         /* mediaItemTransitionReason= */ null);
-    if (isPrepared()) {
+    if (canInitializeLegacyPlaylist(controllerInfo)) {
       initializeLegacyPlaylist();
     }
   }
@@ -766,7 +767,7 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
         /* discontinuityReason= */ null,
         /* mediaItemTransitionReason= */ null);
 
-    if (isPrepared()) {
+    if (hasInitializedLegacyPlaylist) {
       addQueueItems(mediaItems, index);
     }
   }
@@ -820,7 +821,7 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
         /* discontinuityReason= */ null,
         /* mediaItemTransitionReason= */ null);
 
-    if (isPrepared()) {
+    if (hasInitializedLegacyPlaylist) {
       for (int i = fromIndex; i < toIndex && i < legacyPlayerInfo.queue.size(); i++) {
         controllerCompat.removeQueueItem(legacyPlayerInfo.queue.get(i).getDescription());
       }
@@ -888,7 +889,7 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
         /* discontinuityReason= */ null,
         /* mediaItemTransitionReason= */ null);
 
-    if (isPrepared()) {
+    if (hasInitializedLegacyPlaylist) {
       ArrayList<QueueItem> moveItems = new ArrayList<>();
       for (int i = 0; i < (toIndex - fromIndex); i++) {
         moveItems.add(legacyPlayerInfo.queue.get(fromIndex));
@@ -1311,12 +1312,14 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
         /* discontinuityReason= */ null,
         /* mediaItemTransitionReason= */ null);
 
-    if (isPrepared() && hasMedia()) {
+    if (hasInitializedLegacyPlaylist) {
       if (playWhenReady) {
         controllerCompat.getTransportControls().play();
       } else {
         controllerCompat.getTransportControls().pause();
       }
+    } else if (canInitializeLegacyPlaylist(controllerInfo)) {
+      initializeLegacyPlaylist();
     }
   }
 
@@ -1454,17 +1457,16 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
             });
   }
 
-  private boolean isPrepared() {
-    return controllerInfo.playerInfo.playbackState != Player.STATE_IDLE;
-  }
-
-  private boolean hasMedia() {
-    return !controllerInfo.playerInfo.timeline.isEmpty();
+  private static boolean canInitializeLegacyPlaylist(ControllerInfo controllerInfo) {
+    return controllerInfo.playerInfo.playbackState != Player.STATE_IDLE
+        && !controllerInfo.playerInfo.timeline.isEmpty()
+        && (controllerInfo.availablePlayerCommands.contains(Player.COMMAND_PREPARE)
+            || controllerInfo.playerInfo.playWhenReady);
   }
 
   private void initializeLegacyPlaylist() {
+    hasInitializedLegacyPlaylist = true;
     Window window = new Window();
-    checkState(isPrepared() && hasMedia());
     QueueTimeline queueTimeline = (QueueTimeline) controllerInfo.playerInfo.timeline;
     // Set the current item first as these calls are expected to replace the current playlist.
     int currentIndex = controllerInfo.playerInfo.sessionPositionInfo.positionInfo.mediaItemIndex;
@@ -1569,6 +1571,11 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
       List<@NullableType ListenableFuture<Bitmap>> bitmapFutures,
       List<MediaItem> mediaItems,
       int startIndex) {
+    if (released
+        || (controllerCompat.getFlags() & MediaSessionCompat.FLAG_HANDLES_QUEUE_COMMANDS) == 0) {
+      // Stale command, session released or queue support changed.
+      return;
+    }
     for (int i = 0; i < bitmapFutures.size(); i++) {
       @Nullable ListenableFuture<Bitmap> future = bitmapFutures.get(i);
       @Nullable Bitmap bitmap = null;
@@ -1614,7 +1621,7 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
     updateControllerInfo(
         notifyConnected,
         newLegacyPlayerInfo,
-        /* resetPendingLegacyPlayerInfo= */ true,
+        /* hasNewLegacyPlayerInfo= */ true,
         newControllerInfo,
         /* discontinuityReason= */ reasons.first,
         /* mediaItemTransitionReason= */ reasons.second);
@@ -1638,7 +1645,7 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
     updateControllerInfo(
         /* notifyConnected= */ false,
         legacyPlayerInfo,
-        /* resetPendingLegacyPlayerInfo= */ false,
+        /* hasNewLegacyPlayerInfo= */ false,
         newControllerInfo,
         discontinuityReason,
         mediaItemTransitionReason);
@@ -1649,17 +1656,19 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
   private void updateControllerInfo(
       boolean notifyConnected,
       LegacyPlayerInfo newLegacyPlayerInfo,
-      boolean resetPendingLegacyPlayerInfo,
+      boolean hasNewLegacyPlayerInfo,
       ControllerInfo newControllerInfo,
       @Nullable @Player.DiscontinuityReason Integer discontinuityReason,
       @Nullable @Player.MediaItemTransitionReason Integer mediaItemTransitionReason) {
     LegacyPlayerInfo oldLegacyPlayerInfo = legacyPlayerInfo;
     ControllerInfo oldControllerInfo = controllerInfo;
-    if (legacyPlayerInfo != newLegacyPlayerInfo) {
-      legacyPlayerInfo = new LegacyPlayerInfo(newLegacyPlayerInfo);
-    }
-    if (resetPendingLegacyPlayerInfo) {
+    if (hasNewLegacyPlayerInfo) {
+      legacyPlayerInfo = newLegacyPlayerInfo;
       pendingLegacyPlayerInfo = legacyPlayerInfo;
+      if (canInitializeLegacyPlaylist(newControllerInfo)) {
+        // New platform state already has initialized playlist.
+        hasInitializedLegacyPlaylist = true;
+      }
     }
     controllerInfo = newControllerInfo;
 
@@ -2568,17 +2577,6 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
       this.repeatMode = repeatMode;
       this.shuffleMode = shuffleMode;
       this.sessionExtras = sessionExtras != null ? sessionExtras : Bundle.EMPTY;
-    }
-
-    public LegacyPlayerInfo(LegacyPlayerInfo other) {
-      playbackInfoCompat = other.playbackInfoCompat;
-      playbackStateCompat = other.playbackStateCompat;
-      mediaMetadataCompat = other.mediaMetadataCompat;
-      queue = other.queue;
-      queueTitle = other.queueTitle;
-      repeatMode = other.repeatMode;
-      shuffleMode = other.shuffleMode;
-      sessionExtras = other.sessionExtras;
     }
 
     @CheckResult